types: unify plugin hooks type
