+++ packages/compat/webpack/src/types/index.ts
@@ -1,4 +1,3 @@
-export * from './hooks';
 export * from './config';
 export * from './plugin';
 export * from './context';
