+++ packages/core/src/rspack-provider/core/initPlugins.ts
@@ -67,5 +67,7 @@ export function getPluginAPI({
     onBeforeCreateCompiler: hooks.onBeforeCreateCompilerHook.tap,
     onAfterStartDevServer: hooks.onAfterStartDevServerHook.tap,
     onBeforeStartDevServer: hooks.onBeforeStartDevServerHook.tap,
+    modifyWebpackChain: () => {},
+    modifyWebpackConfig: () => {},
   };
 }
