+++ packages/core/src/rspack-provider/types/index.ts
@@ -1,4 +1,3 @@
-export * from './hooks';
 export * from './plugin';
 export * from './context';
