+++ packages/shared/src/config.ts
@@ -18,7 +18,6 @@ import {
 import type {
   RsbuildTarget,
   BundlerChainRule,
-  HtmlConfig,
   RsbuildConfig,
   InspectConfigOptions,
   CreateRsbuildOptions,
@@ -39,7 +38,6 @@ import type { minify } from 'terser';
 import fs from 'fs-extra';
 
 import _ from 'lodash';
-import { isProd } from './utils';
 import { DEFAULT_DEV_HOST } from './constants';
 import { getJSMinifyOptions } from './minimize';
